+++ public/app/angular/components/info_popover.ts
