+++ public/app/angular/components/sql_part/sql_part.ts
