+++ public/app/angular/components/switch.ts
