+++ public/app/angular/services/timer.ts
