+++ public/app/core/services/all.ts
@@ -1,4 +1,3 @@
 import './util_srv';
 import './context_srv';
-import './timer';
 import './backend_srv';
