+++ public/app/plugins/datasource/mysql/sql_part.ts
@@ -1,4 +1,4 @@
-import { SqlPartDef, SqlPart } from 'app/core/components/sql_part/sql_part';
+import { SqlPartDef, SqlPart } from 'app/angular/components/sql_part/sql_part';
 
 const index: any[] = [];
